chore(test): consolidate all test WASM files
