+++ e2e/e2e_register_core_contract_test.go
