+++ testutil/testwasms/core_contract.wasm
