+++ testutil/testwasms/sample_tally.wasm
