+++ testutil/testwasms/sample_tally_2.wasm
