+++ x/tally/keeper/filter_and_tally_test.go
@@ -9,8 +9,8 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/stretchr/testify/require"
 
+	"github.com/sedaprotocol/seda-chain/testutil/testwasms"
 	dataproxytypes "github.com/sedaprotocol/seda-chain/x/data-proxy/types"
-	"github.com/sedaprotocol/seda-chain/x/tally/keeper/testdata"
 	"github.com/sedaprotocol/seda-chain/x/tally/types"
 	wasmstoragetypes "github.com/sedaprotocol/seda-chain/x/wasm-storage/types"
 )
@@ -178,7 +178,7 @@ func TestExecutorPayout(t *testing.T) {
 	err := f.tallyKeeper.SetParams(f.Context(), defaultParams)
 	require.NoError(t, err)
 
-	tallyProgram := wasmstoragetypes.NewOracleProgram(testdata.SampleTallyWasm2(), f.Context().BlockTime())
+	tallyProgram := wasmstoragetypes.NewOracleProgram(testwasms.SampleTallyWasm2(), f.Context().BlockTime())
 	err = f.wasmStorageKeeper.OracleProgram.Set(f.Context(), tallyProgram.Hash, tallyProgram)
 	require.NoError(t, err)
